fix: apply cargo fmt formatting
