+++ lago-client/Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT"
 repository = "https://github.com/getlago/lago-rust-client"
 
 [dependencies]
-lago-types = { version = "0.1.21", path = "../lago-types" }
+lago-types = "0.1.20"
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
